Unit test table component
